Overlap Bug
Enemies are not overlapping while chasing. Speed is now double

+++ src/core/src/com/mygdx/game/sprites/enemies/Enemy.java
@@ -8,7 +8,7 @@ public abstract class Enemy extends Character{
     private int spawnRate;
     private int coolDown;
 
-    public Enemy(int health, int maxHealth, int speed, int damage, int spawnRate, int coolDown){
+    public Enemy(int health, int maxHealth, double speed, int damage, int spawnRate, int coolDown){
         super(health, maxHealth, speed, damage);
         this.spawnRate = spawnRate;
         this.coolDown = coolDown;
@@ -36,7 +36,7 @@ public abstract class Enemy extends Character{
         float yDifference = yLocation - getPosition().y;
         Vector2 distance = new Vector2(xDifference, yDifference);
         Vector2 unitVector = new Vector2(xDifference / distance.len(), yDifference / distance.len());
-        Vector2 newVelocity = unitVector.scl(speed);
+        Vector2 newVelocity = unitVector.scl((float) speed);
 
         setVelocity(newVelocity);
     }
