+++ src/core/src/com/mygdx/game/sprites/enemies/Grunt.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.math.Rectangle;
 
 public class Grunt extends Enemy {
-    public Grunt(int health, int maxHealth, int speed, int damage, int spawnRate, int coolDown){
+    public Grunt(int health, int maxHealth, double speed, int damage, int spawnRate, int coolDown){
         super(health, maxHealth, speed, damage, spawnRate, coolDown);
         elementTexture = new Texture("paladin.png");
     }
